requirements.txt
Django==5.2.6
djangorestframework==3.16
Pillow==11.3.0
autoslug==1.0.6
django-sortedm2m==4.0.0
djangorestframework-simplejwt==5.3.1
django-autoslug-1.8.0
django-filter==25.1

